Avoid infinite loops when the line is incorrect
This seems to solves #140 however we need clean tests for that function, it is not clear why it is written that way and what kind of cases it is trying to solve. Having a non conditional loop requires that we make absolutely sure that it can exit the loop. 

Maybe in the mean time we need to add a counter that is initialized with the length of the line and make it spit a warning if it gets there. 
At least that would not make it crash and we would have debug information on why it crashed.