Changed the definition of propertyListNotEmpty.
The new definition uses 'verbPath or verbSimple' instead of 'verb'. Note that 'verb' is also defined in 'verbPath or verbSimple'. The need for change came from a query that included nested bnode syntax and a property path sequence '/'; As soon as the prop.path sequence sign is inserted, the specific query line is marked as invalid. However, the query runs and evaluates just fine.
verbPath handles property paths, hence the change.

NOTE: 'verb' is only used to define propertyListNotEmpty. Should you decide that the change is appropriate, you might consider omitting 'verb' overall, as it will be obsolete.

Eetz